cart icon badge works
